Fixes #145 - Rename package in opentracing-noop to io.opentracing.noop
