+++ opentracing-noop/src/main/java/io/opentracing/noop/NoopActiveSpanSource.java
@@ -11,7 +11,12 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing;
+package io.opentracing.noop;
+
+import io.opentracing.ActiveSpan;
+import io.opentracing.ActiveSpanSource;
+import io.opentracing.Span;
+import io.opentracing.SpanContext;
 
 import java.util.Map;
