+++ opentracing-noop/src/main/java/io/opentracing/noop/NoopSpanBuilder.java
@@ -11,7 +11,13 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing;
+package io.opentracing.noop;
+
+import io.opentracing.ActiveSpan;
+import io.opentracing.BaseSpan;
+import io.opentracing.Span;
+import io.opentracing.SpanContext;
+import io.opentracing.Tracer;
 
 import java.util.Collections;
 import java.util.Map;
