+++ opentracing-noop/src/main/java/io/opentracing/noop/NoopSpanContext.java
@@ -11,7 +11,9 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing;
+package io.opentracing.noop;
+
+import io.opentracing.SpanContext;
 
 import java.util.Collections;
 import java.util.Map;
