+++ opentracing-noop/src/main/java/io/opentracing/noop/NoopTracerFactory.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing;
+package io.opentracing.noop;
 
 public final class NoopTracerFactory {
